% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{adjusted_risk}
\alias{adjusted_risk}
\title{adjusted_risk}
\usage{
adjusted_risk(prop, RR)
}
\arguments{
\item{prop}{A vector of proportions of the population that belong
to each URG (Unit risk group)}

\item{RR}{A vector of the relative risks of for each URG. The
first of these is the referent group and therefore must be
equal to 1}
}
\value{
A vector of Adjusted risks
}
\description{
Adjusted Risk
}
\details{
Calculate the adjusted risk for each of the unit risk groups
(URG). This can be used at both the herd and the animal level. The
proportion vector, for herd level, is therefore the proportion
herds in the population that are in each of the unit risk
groups. The proportion vector for animal level is the proportion
of animals within a given herd that are in each URG.
}
\author{
Thomas Rosendal
}
