% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{EffProbInf}
\alias{EffProbInf}
\title{EffProbInf}
\usage{
EffProbInf(dp, AR)
}
\arguments{
\item{dp}{A vector The design prevalence}

\item{AR}{A vector of the adjusted risks of the unit risk groups}
}
\value{
A vector of EPI
}
\description{
EffProbInf
}
\details{
Calculate the effective probability of infection (EPI) for each unsit
risk group in the population. This could be either at the herd
level or within herd level. The dp for herds is therefore the
minimum prevlance among herds that you would like to design the
surveillance system to be able to detect. The dp for within herds
is therefore the minimum prevalance of the disease within a herd
among the animals that you would like to design the surveillance
system to detect.
}
\author{
Thomas Rosendal
}
