% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{equivalant}
\alias{equivalant}
\title{equivalent}
\usage{
equivalant(x, y, tolerance = 1e-07)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{tolerance}{a tolerance value}
}
\value{
logical
}
\description{
A function used to check if two values are equilavent to an
accepted tolerance
}
