% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_se.R
\name{hse_infinite}
\alias{hse_infinite}
\title{hse_infinite}
\usage{
hse_infinite(n_tested, test_Se, dp)
}
\arguments{
\item{n_tested}{The number tested in each URG}

\item{test_Se}{The sensitivity of the test (length = 1). If you
have reason to believe that the test sensitivity is different
for different URG. Then supply a vector of Sensitivities. This
could conceivably be because of using different tests for
different samples from different URG.}

\item{dp}{The is a vector (length 1) of the the design prevalance
(df) in the case where there is only one unit risk group (URG)
in the herd. Or a vector (length n) of EPIn for each of the
URG in the herd.}
}
\value{
A vector (length 1)
}
\description{
Herd Sensitivity calculated with the assumption of an infinite population
}
\details{
Calculate the Herd sensitivity when multiple samples from
individual units within the herd. The function does not use the
population size to adjust the estimate. This is consistent with
the assumption of an infinite population size and is generally
used when less than 10% of the population has been sampled.
}
\author{
Thomas Rosendal
}

