% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_discounting.R
\name{post_fr}
\alias{post_fr}
\title{post_fr}
\usage{
post_fr(prior_fr, Se)
}
\arguments{
\item{prior_fr}{The prior probability of freedom}

\item{Se}{The sensitivity of the surveillance system}
}
\value{
A vector
}
\description{
Calculate the posterior probability of freedom from the prior and
the sensitivity of the system
}
\details{
The prior probabailty of freedom at the beginning of the
surveillance initiative is a value that is based on some external
evidence. Often 0.5 is used as a conservative estimate of the
probabaility that the population is free from the disease. For
subsequent time intervals in the surveillance system, the prior
year's posterior probability of freedom is used (plus a risk of
introduction) as the prior probability in this calculation.
}
\examples{
## Calculate the posterior propability of freedom after applying a
#sensitivity to a prior probability of freedom:
post_pf <- post_fr(0.5, 0.4)
}
