% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_discounting.R
\name{prior_fr}
\alias{prior_fr}
\title{prior_fr}
\usage{
prior_fr(post_fr, intro)
}
\arguments{
\item{post_fr}{The posterior probability of freedom (year = k-1)}

\item{intro}{The annual probability of introduction}
}
\value{
A vector (length = 1) The prior probability of freedom (year = k)
}
\description{
Calculate the prior probability of freedom (year = k)
}
\details{
In order to calculate the posterior probability of freedom (year =
k) , the prior probability of freedom (year = k) is first
calculated from the posterior probability of freedom (year = k-1)
from the previous year and the annual probability that the disease
is introduced into the population.
}
\author{
Thomas Rosendal
}

