% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{rpert}
\alias{rpert}
\title{rpert}
\usage{
rpert(n, x.min, x.max, x.mode, lambda = 4)
}
\arguments{
\item{n}{number of samples}

\item{x.min}{The minimum value in the sample}

\item{x.max}{The maximum value in the sample}

\item{x.mode}{The mode of the sample}

\item{lambda}{lambda}
}
\value{
a numeric vector of length n
}
\description{
Sample a pert distribution
}
\details{
Returns samples from a pert distribution
}
