% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_se.R
\name{sysse}
\alias{sysse}
\title{sysse}
\usage{
sysse(dp, hse)
}
\arguments{
\item{dp}{The vector of EPIH for all herds tested in the surveillance system}

\item{hse}{The calculated hse for all the herds tested in teh surveillance system}
}
\value{
A vector (length 1)
}
\description{
Calculate the surveillance system sensitivity
}
\details{
Takes a vector of the sensitivity of herds tested in the
surveillance system and a vector of the effective probability of
infection in the herds (EPIH) to calculate the total surveillance system
sensitvity for the entire program.
}
\author{
Thomas Rosendal
}

