% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_se.R
\name{sysse_finite}
\alias{sysse_finite}
\title{sysse}
\usage{
sysse_finite(dp, hse, N)
}
\arguments{
\item{dp}{The vector of EPIH for all herds tested in the
surveillance system.}

\item{hse}{The calculated hse for all the herds tested in the
surveillance system.}

\item{N}{The total number of herds in the population.}
}
\value{
A vector (length 1)
}
\description{
Calculate the surveillance system sensitivity for a finite population of herds
}
\details{
Takes a vector of the sensitivity of herds tested in the
surveillance system and a vector of the effective probability of
infection in the herds (EPIH) to calculate the total surveillance
system sensitivity for the entire program. This is adjusted for
the the total number of herds in the population.
}
\author{
Thomas Rosendal
}

