% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{valid_proportions}
\alias{valid_proportions}
\title{valid_proportions}
\usage{
valid_proportions(x, tolerance = 1e-07)
}
\arguments{
\item{x}{numeric}

\item{tolerance}{a tolerance value}
}
\value{
logical
}
\description{
A function used to check if a vector of proportions is valid
}
