% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{rpert}
\alias{rpert}
\title{rpert}
\usage{
rpert(n, x.min, x.max, x.mode, lambda = 4)
}
\arguments{
\item{n}{number of samples}

\item{x.min}{The minimum value in the sample}

\item{x.max}{The maximum value in the sample}

\item{x.mode}{The mode of the sample}

\item{lambda}{lambda}
}
\value{
a numeric vector of length n
}
\description{
Sample a pert distribution
}
\details{
Returns samples from a pert distribution
}
\examples{
## Generate 10000 samples from a pert distribution with a minimum
## of 2, a max of 5, and a mode of 4.
samples <- rpert(10000, 2, 5, 4)
hist(samples)

## Generate a
}
